Match file paths recursively
In some projects, we nest Rails engines or Ruby gems inside the project
directory. By prefixing the file paths with wildcards, the rules
defined in the configuration file are applied even in these nested
directories. This change enables us to use the style gem in projects
that nest engines or gems.