Ensure that versions are always in sync
When cutting a new release, three different versions need to be updated:

- The gemspec has the definitive version of the gem in it.
- The CHANGELOG contains the changes made in the different versions.
- The README file includes the current version in the installation
  instructions.

Not every developer is aware that these three places need to be changed,
and even then it is easy to miss one or two. To protect us against
inconsistencies, specs have been added that check that these three
versions exist, are in sync and meet certain requirements. The tests are
automatically run on CircleCI for each pull request to ensure that the
master branch is in a healthy state.